fix: revert to simpler and more dry approach
