+++ pages/series/index.vue
@@ -145,20 +145,22 @@ const { isLoading, isMobile, hasMore, desktopItemList, mobileItemList, totalPage
 // Element reference for the scroll target
 const resultsSection = ref(null)
 // usePaginationScroll composable
+const { scrollTo } = usePaginationScroll()
 
-usePaginationScroll(resultsSection, {
-  isMobile,
-  hasResults: computed(() => parsedEventSeries.value.length > 0),
-  offset: 300,
-  onPageChange: async () => {
-    isLoading.value = false
-    currentPage.value = route.query.page ? parseInt(route.query.page) : 1
-    isMobile.value ? mobileItemList.value = [] : desktopItemList.value = []
-    hasMore.value = true
+watch(() => route.query, async (newVal, oldVal) => {
+  isLoading.value = false
+  currentPage.value = route.query.page ? parseInt(route.query.page) : 1
+  isMobile.value ? mobileItemList.value = [] : desktopItemList.value = []
+  hasMore.value = true
 
-    await searchES()
-  },
-})
+  await searchES()
+  // Restore scroll position
+  // // Scroll after DOM updates
+  await nextTick()
+  if (!isMobile.value && route.query.page && resultsSection.value && parsedEventSeries.value.length > 0) {
+    await scrollTo(resultsSection)
+  }
+}, { deep: true, immediate: true })
 
 const parseViewSelection = computed(() => {
   return currentView.value === 'current' ? 1 : 0
